fix(Terminal): fix the client binding connexion
note: this took me some hours to debug, the error was that, at line 448 of the previous file version, I accidentaly erase the consoleSocket variable (which contain the socket instance for the console) with the web socket instance . In fact, if the getConsoleSocket function worked it would not be a issue, as I would be getting the right variable. So because of  a syntax error in getConsoleSocket (missing parentheses) the consoleSocket would be in fact  refering as the web client and the terminal-output, terminal-resize and terminal-input would be sent instead to the web. This was very tricky to debug because the issue wasn't showing up every time and this is because the getConsoleSocket function will always return the first client connected so this explain the fact that the issue only show up when the overlay connect to the server AFTER the web client!! With the right analysis, everything can be explained, wonderful!